+++ combinators/sequencing/mod.ts
@@ -5,4 +5,8 @@
  */
 
 export { between } from "./between.ts";
+export { foldL } from "./foldL.ts";
+export { foldL1 } from "./foldL1.ts";
+export { foldR } from "./foldR.ts";
+export { foldR1 } from "./foldR1.ts";
 export { seq } from "./seq.ts";
